% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comp.R
\docType{package}
\name{model.comp}
\alias{model.comp}
\title{model.comp: a package for model comparisons}
\description{
The model.comp package currently provides the compare_models() function which
fits specified models and evaluates specified performance metrics for a given
dataset.
}
