% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_models}
\alias{compare_models}
\title{Compare some performance measures of models fitted to your data}
\usage{
compare_models(
  data,
  eqn,
  models = c("logreg"),
  outputs = c("fitted_model", "fitted_values")
)
}
\arguments{
\item{data}{data.frame containing at least the response and all explanatory variables in your equation}

\item{eqn}{string specifying the formula for the models to fit}

\item{models}{vector/list of the names of models to compare - currently supports logistic regression
(logreg), classification and regression trees (cart), and pruned classification/regression trees (prune)}

\item{outputs}{vector/list specifying the types of outputs that should be returned - currently supports
the model fitted (fitted_model), and the fitted/predicted values for the data given, calculated with
the model fitted (fitted_values)}
}
\value{
returns a data.frame with a row for each model fitted and columns for each output requested
}
\description{
Compare some performance measures of models fitted to your data
}
\examples{
compare_models(data = gapminder::gapminder, eqn = "lifeExp ~ gdpPercap", models = c('logreg'))
}
